removed unused pages
