--- src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/ChooseOrganisationPage.cs
@@ -1,19 +1,0 @@
-﻿using SFA.DAS.Registration.UITests.Project;
-using TechTalk.SpecFlow;
-
-namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
-{
-    public class ChooseOrganisationPage : EIBasePage
-    {
-        protected override string PageTitle => "Choose organisation";
-
-        public ChooseOrganisationPage(ScenarioContext context) : base(context)  { }
-
-        public EIHubPage SelectFirstEntityInChooseOrgPageAndContinue()
-        {
-            formCompletionHelper.SelectRadioOptionByText(objectContext.GetOrganisationName());
-            formCompletionHelper.Click(ContinueButton);
-            return new EIHubPage(context);
-        }
-    }
-}
