+++ src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/ConfirmApprenticeCancelPage.cs
@@ -16,6 +16,5 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
             Continue();
             return new ApplicationCancelledPage(context);
         }
-
     }
 }
