+++ src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/EIHubPage.cs
@@ -1,5 +1,4 @@
-﻿using SFA.DAS.Registration.UITests.Project.Tests.Pages;
-using TechTalk.SpecFlow;
+﻿using TechTalk.SpecFlow;
 
 namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
 {
@@ -8,12 +7,6 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
         protected override string PageTitle => "Hire a new apprentice payment";
 
         public EIHubPage(ScenarioContext context) : base(context)  { }
-
-        public EIBeforeYouStartPage ClickApplyLinkOnEIHubPage()
-        {
-            HireANewApprenticePayment();
-            return new EIBeforeYouStartPage(context);
-        }
 
         public ViewApplicationsPage NavigateToEIViewApplicationsPage()
         {
@@ -38,15 +31,6 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
             formCompletionHelper.ClickLinkByText("Change organisation and finance details");
             return new ChangeBankDetailsPage(context);
         }
-
-        public HomePage NavigateToHomePage()
-        {
-            ClickBackLink();
-
-            return new HomePage(context);
-        }
-
-        private void HireANewApprenticePayment() => formCompletionHelper.ClickLinkByText("Apply for the hire a new apprentice payment");
 
         private void ViewApplications() => formCompletionHelper.ClickLinkByText("View applications");
     }
