--- src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/NewLegalAgreementRequiredShutterPage.cs
@@ -1,22 +1,0 @@
-﻿using OpenQA.Selenium;
-using SFA.DAS.Registration.UITests.Project;
-using SFA.DAS.Registration.UITests.Project.Tests.Pages;
-using TechTalk.SpecFlow;
-
-namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
-{
-    public class NewLegalAgreementRequiredShutterPage : EIBasePage
-    {
-        protected override string PageTitle => $"{ObjectContextExtension.GetOrganisationName(objectContext)} needs to accept a new employer agreement";
-
-        public NewLegalAgreementRequiredShutterPage(ScenarioContext context) : base(context)  { }
-
-        public By ViewAgreement => By.ClassName("govuk-button");
-
-        public YourOrganisationsAndAgreementsPage ViewLegalAgreement()
-        {
-            formCompletionHelper.ClickButtonByText(ViewAgreement, "View agreement");
-            return new YourOrganisationsAndAgreementsPage(context);
-        }
-    }
-}
