--- src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/NotEligibleShutterPage.cs
@@ -1,45 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using OpenQA.Selenium;
-using TechTalk.SpecFlow;
-
-namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages
-{
-    public class NotEligibleShutterPage : EIBasePage
-    {
-        protected override string PageTitle => $"One or more apprentices are not eligible for the payment";
-
-        #region Locators
-        private readonly ScenarioContext _context;
-
-        private By IneligibleApprenticesList => By.CssSelector("[data-name=ineligible-apprenticeship]");
-        private By EmploymentStartDateFields => By.Name("ineligible-apprenticeship-start-date");
-        private By CancelApplicationButton => By.Id("cancelApplication");
-        private By ContinueApplicationButton => By.Id("continueApplication");
-        #endregion
-
-        public NotEligibleShutterPage(ScenarioContext context, bool verifypage = true) : base(context, verifypage)
-        {
-            _context = context;
-        }
-
-        public int NumberOfIneligibleApprenticeships => pageInteractionHelper.FindElements(IneligibleApprenticesList).Count;
-        public bool CancelApplicationButtonExists => pageInteractionHelper.IsElementPresent(CancelApplicationButton);
-        public bool ContinueApplicationButtonExists => pageInteractionHelper.IsElementPresent(ContinueApplicationButton);
-
-        public List<DateTime> EmploymentStartDates
-        {
-            get
-            {
-                var startDates = new List<DateTime>();
-                var startDateElements = pageInteractionHelper.FindElements(EmploymentStartDateFields);
-                foreach (var startDateElement in startDateElements)
-                {
-                    startDates.Add(DateTime.Parse(startDateElement.Text));
-                }
-
-                return startDates;
-            }
-        }
-    }
-}
