+++ src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/VRF/VRFAmendmentsTabPage.cs
@@ -8,7 +8,6 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages.VRF
         protected override string PageTitle => "Provide amendments";
 
         #region Locators
-        
         private By RemittanceEmailCheckBox => By.Id("c_rem_email");
         private By RemittanceEmailCheckBoxSelectionStatus => By.XPath("//fieldset[@id='c_rem_email']/following-sibling::i[@title='Field is valid']");
         private By ChangeBankDetailsRadioSelectionStatus => By.XPath("//fieldset[@id='change_bank']/following-sibling::i[@title='Field is valid']");
