+++ src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/VRF/VRFIntroductionTabPage.cs
@@ -10,7 +10,6 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages.VRF
         protected override string PageTitle => "Provide organisation information about your banking and payments to DfE";
 
         #region Locators
-        
         private By VRFCookieCloseButton => By.Id("close-cookie-message");
         protected override By PageHeader => By.CssSelector("h1");
         #endregion
