+++ src/SFA.DAS.EmployerIncentives.UITests/Project/Tests/Pages/VRF/VRFReceivedDetailsConfirmPage.cs
@@ -10,18 +10,10 @@ namespace SFA.DAS.EmployerIncentives.UITests.Project.Tests.Pages.VRF
         protected override By PageHeader => By.XPath("//h1[text()='We have received your details']");
 
         #region Locators
-        
-        private By CaseIdSummary => By.CssSelector(".panel-gds p");
         private By ReturnToEasLink => By.CssSelector(".submission-message a");
         #endregion
 
         public VRFReceivedDetailsConfirmPage(ScenarioContext context) : base(context, false) => frameHelper.SwitchFrameAndAction(() => VerifyPage());
-
-        public ApplicationCompletePage ReturnToEasApplicationCompletePage()
-        {
-            ReturnToEAS();
-            return new ApplicationCompletePage(context);
-        }
 
         public ApplicationClosedPage ReturnToApplicationClosedPage()
         {
